m22_queues/hw22/requirements.txt
Flask
Celery
Flask-Mail
redis
pillow
flower